replace go.rice with go:embed
